Move scoreboard buttons to separate components
