+++ src/components/questions/buzzer/BuzzerQuestionDialog.vue
